Work around overloaded puppetmasters
That is, everytime a run exits with a non-zero exit code (as handled by
puppetgo, so puppet exit 2 -> 0), the next run starts with a delay of up
to 240 seconds.

That should spread out the load of repeated tries to run that overload
the (already loaded) puppetmaster machine, as such helping the whole
plethora of servers finally get their run done...

Not nice, but works.